// app/layout/models/department.model.ts
export class Department {
  noOfItems: number;
  departmentID: number;
  companyID: number;
  companyName: string;
  departmentDescription: string;
  departmentTypeID: number;
  deleteObject: string;
  activeFlag: boolean;
  displayPromptMethodID: number;
  audibleAgeVerificationAlertFlag: boolean;
  isModified: boolean;
  isStoreLocationDeleted: boolean;
  isDepartmentOpen: boolean;
  isFractionalQtyAllowedFlag: boolean;
  isLoyaltyRedeemEligibleFlag: true;
  isItemReturnableFlag: boolean;
  isPriceRequiredFlag: boolean;
  areSpecialDiscountsAllowedFlag: boolean;
  departmentProfitMargin: number;
  profitPercent: number;
  allowFoodStampsFlag: boolean;
  minimumOpenSaleAmount: number;
  maximumOpenSaleAmount: number;
  displayPromptMethodDesc: string;
  departmentTypeName: string;
  departmentLocationID: number;
  storeLocationID: number;
  posDepartmentCode: string;
  posSyncStatusID: number;
  posSyncStatusCode: string;
  storeLocationTaxID: number;
  taxStrategyDescription: string;
  storeLocationSalesRestrictionID: number;
  salesRestrictionDescription: string;
  posDepartmentDescription: string;
  currentAsOfDateTime: Date;
  updateDescInEDIInvoiceFlag: boolean;
  updateSellingPriceInEDIInvoiceFlag: boolean;
  salesRestrictionRequiredFlag: boolean;
  isDepartmentNegative: boolean;
  lastSyncDateTime: Date;
  storeName: string;
  isBlueLaw1Enabled: boolean;
  isBlueLaw2Enabled: boolean;
  createdDateTime: Date;
  lastModifiedBy: string;
  lastModifiedDateTime: Date;
  naxmlDepartmentTypeCodeID: number;
  priceRequiredFlag: boolean;
  isInsideSalesDept: boolean;
  displayOrder: number;
  srNo: number;
  noOfItemsInDepartment: number;
}